minor tweaks for better compatibility with the AMD/ATI drivers
git-svn-id: http://code.icecube.wisc.edu/svn/projects/clsim/trunk@97732 16731396-06f5-0310-8873-f7f720988828

+++ private/clsim/function/I3CLSimScalarFieldIceTiltZShift.cxx
@@ -124,9 +124,9 @@ double I3CLSimScalarFieldIceTiltZShift::GetValue(double x, double y, double z) c
 
     const double nr = lnx*x+lny*y;
 
-    for(int j=1; j<distancesFromOriginAlongTilt_.size(); j++) 
+    for(int j=1; j<static_cast<int>(distancesFromOriginAlongTilt_.size()); j++) 
     {
-        if((nr<distancesFromOriginAlongTilt_[j]) || (j==distancesFromOriginAlongTilt_.size()-1))
+        if((nr<distancesFromOriginAlongTilt_[j]) || (j==static_cast<int>(distancesFromOriginAlongTilt_.size())-1))
         {
             const double thisDistanceBinWidth = (distancesFromOriginAlongTilt_[j] - distancesFromOriginAlongTilt_[j-1]);
